refactor: Move data preprocessing to a separate script in the model directory
